ATR-654: created a base class for nodes that represent IsActive graph methods
